Move all /workflow to workflow service, update subscribe signatures - #4994
