+++ ui/app/src/services/publishing.ts
@@ -19,7 +19,7 @@ import { Observable } from 'rxjs';
 import { map, mapTo, pluck } from 'rxjs/operators';
 import { LegacyItem } from '../models/Item';
 import { pluckProps, toQueryString } from '../utils/object';
-import { PublishingParams, PublishingStatus, PublishingTarget } from '../models/Publishing';
+import { PublishingStatus, PublishingTarget } from '../models/Publishing';
 import { Api2BulkResponseFormat, Api2ResponseFormat } from '../models/ApiResponse';
 import { PagedArray } from '../models/PagedArray';
 
@@ -84,35 +84,6 @@ export interface GoLiveResponse {
   invalidateCache: boolean;
   success: boolean;
   message: string;
-}
-
-export function publish(siteId: string, data: PublishingParams): Observable<boolean> {
-  return postJSON('/studio/api/2/workflow/publish', {
-    siteId,
-    ...data
-  }).pipe(mapTo(true));
-}
-
-export function requestPublish(siteId: string, data: PublishingParams): Observable<boolean> {
-  return postJSON('/studio/api/2/workflow/request_publish', {
-    siteId,
-    ...data
-  }).pipe(mapTo(true));
-}
-
-export function approve(siteId: string, data: PublishingParams): Observable<boolean> {
-  return postJSON('/studio/api/2/workflow/approve', {
-    siteId,
-    ...data
-  }).pipe(mapTo(true));
-}
-
-export function reject(siteId: string, items: string[], comment: string): Observable<boolean> {
-  return postJSON('/studio/api/2/workflow/reject', {
-    siteId,
-    items,
-    comment
-  }).pipe(mapTo(true));
 }
 
 export function fetchStatus(siteId: string): Observable<PublishingStatus> {
